requirements.txt
jupyterlab
tqdm
aiohttp
pandas
python-dotenv
httpx
requests
pytest
beautifulsoup4
